# backend/docs/swagger.yaml
definitions:
...
  dto.CreateUserInputDTO:
    properties:
      biological_sex:
        type: string
      birth_date:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateUserInputDTO:
    properties:
      biological_sex:
        type: string
      birth_date:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
...
  entity.User:
    properties:
      biologicalSex:
        type: string
      birthDate:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user:
    get:
      description: Retrieves a list of all registered user profiles sorted by name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Creates a user
      parameters:
      - description: User Profile Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user profile
      tags:
      - user
  /user/{id}:
    delete:
      description: Permanently removes a user and their associated data from the system.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user profile
      tags:
      - user
    get:
      description: Retrieves the basic profile information of a specific user.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user profile by ID
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Updates the name, email, or biological data of an existing user.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid update data
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user profile
      tags:
      - user
swagger: "2.0"